% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{alpha}
\alias{alpha}
\title{alpha}
\usage{
alpha(col, alpha)
}
\arguments{
\item{col}{RGB color}
}
\description{
Give transparency value to colors

Takes provided colors and gives them the specified alpha (ie transparency) value
}
\author{
Marcin Imielinski
}
\keyword{internal}
