% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventCallers.R
\name{amp}
\alias{amp}
\title{Amplifications}
\usage{
amp(
  gg,
  jcn.thresh = 8,
  cn.thresh = 2,
  fbi.cn.thresh = 0.5,
  n.jun.high.bfb.thresh = 26,
  n.jun.high.dm.thresh = 31,
  width.thresh = 1e+05,
  fbi.width.thresh = 1e+05,
  mc.cores = 1,
  mark = TRUE,
  mark.col = "purple",
  mark.nos = FALSE,
  min.nodes = 3,
  min.jun = 2
)
}
\arguments{
\item{gg}{gGraph}

\item{jcn.thresh}{minimal ALT edge junction threshold to classify a high copy cluster.
Default: 9}

\item{cn.thresh}{minimal node copy number in ploidy units to classify a high copy cluster.
Default: 2}

\item{fbi.cn.thresh}{fraction of total CN in cluster that is contributed to by
fold back inversions, if higher than this will call a BFB.
Default: 0.5}

\item{n.jun.high.bfb.thresh}{max number of high copy junctions in a fbi.cn
high cluster, if fbi.cn is high and high copy junctions exceed this, then will call a tyfonas.
Default: 26}

\item{n.jun.high.dm.thresh}{double minute threshold set for junctions. If fbi.cn
is low, and high copy junctions exceed this thresh, it will call it cpxdm, else
dm.
Default: 31}

\item{width.thresh}{minimum width to consider for an amplification event.
Default: 1e5}

\item{mark.nos}{(logical) Default: FALSE}

\item{min.nodes}{(numeric) minimum number of nodes for a cluster to be
designated amp-NOS. Default: 3}

\item{min.jun}{(numeric) minimum number of aberrant junctions for a cluster
to be designated amp-NOS. Default: 2}
}
\value{
gg of amplification events found.
}
\description{
Classifies high-level amplifications in copy number annotated gGraph.  Default parameters
should be used.
}
\details{
Amplification events are defined into 3 groups, bfb, tyfonas, and dm events.
More details can be found in the tutorial:
\href{http://mskilab.com/gGnome/tutorial.html#Complex_amplicons_(bfb,_dm,_tyfonas)}{Complex Amplicons}
}
