% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_by_field_from_seqnames}
\alias{append_by_field_from_seqnames}
\title{Appending by field in seqnames to metadata column}
\usage{
append_by_field_from_seqnames(x, by = NULL)
}
\value{
GRanges with seqnames keyed with by field with metadata column(s) appended.
}
\description{
For use downstream of gr_construct_by.
By field that is keyed onto the original seqnames
can be rescued and re-appended to a GRanges output.
For example if you do GenomicRanges set operations like
in GenomicRanges::setdiff(), ::union(), or ::intersect(), 
the metadata fields will be stripped. But the by field
can be re appended as a column.
}
\author{
Kevin Hadi
}
