% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{bcheck}
\alias{bcheck}
\title{bcheck}
\usage{
bcheck(gg)
}
\arguments{
\item{gg}{gGraph with node and edge fields 'cn', node field loose.cn.left and loose.cn.right}
}
\description{
Checks if genome graph with node and edge cn's annotated and loose cn's is balanced.
}
