% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\docType{class}
\name{bGraph-class}
\alias{bGraph-class}
\title{bGraph: junction-balanced graph
Descendant of gGraph class, where junction balance restraint must be met at all times.}
\description{
bGraph: junction-balanced graph
Descendant of gGraph class, where junction balance restraint must be met at all times.
}
