% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{binstats}
\alias{binstats}
\title{binstats}
\usage{
binstats(
  gg,
  bins,
  by = NULL,
  field = NULL,
  purity = gg$meta$purity,
  ploidy = gg$meta$ploidy,
  loess = TRUE,
  min.bins = 3,
  verbose = TRUE,
  min.var = 0.1
)
}
\arguments{
\item{gg}{gGraph}

\item{bins}{GRanges with field $cn or field field}

\item{by}{optional character vector specifying metadata field(s) shared by gg$nodes and bins to which bin / node overlaps will be limited to}

\item{field}{character vector field of bins to convert to (NULL)}

\item{purity}{purity parameter either specified together with field or embedded in gg$meta, must be specified if field is not NULL}

\item{ploidy}{ploidy parameter either specified together with field or embedded in gg$meta, must be specified if field is not NULL}

\item{loess}{logical flag whether to smooth / fit variance using loess (FALSE)}

\item{min.bins}{minimum number of bins to use for intra segment variance computation (3)}

\item{min.var}{minimal allowable per segment bin variance, which will ignore segments with very low variance due to all 0 or other reasons (0.1)}
}
\value{
gGraph whose nodes are annotated with $cn and $weight field
}
\description{
Given GRanges of binned eg read depth data with field $cn, crosses 
nodes in graph and annotates graph nodes with
'cn' and 'weight'.  Done upstream of balance. 

If "by" field(s) specified and these fields exist in both
bins and graph nodes, then will use these in the overlaps query

If field, purity, and ploidy provided then will
also transform read depth data in bin column "field"
using purity and ploidy to generate
}
\author{
Marcin Imielinski
}
