% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{booth_rotate}
\alias{booth_rotate}
\title{Use Booth's algorithm to disambiguate a circular vector}
\usage{
booth_rotate(x)
}
\arguments{
\item{x}{vector input}
}
\value{
x_rot rotated vector
}
\description{
Use Booth's algorithm to disambiguate a circular vector
}
