% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsUtils.R
\name{brewer.master}
\alias{brewer.master}
\title{brewer.master}
\usage{
brewer.master(n, palette = NULL, wes = FALSE, list = FALSE)
}
\arguments{
\item{n}{TODO}

\item{palette}{character specifyign pallette to start with (options are: Blues, BuGn, BuPu, GnBu, Greens Greys, Oranges, OrRd, PuBu, PuBuGn, PuRd, Purples, RdPu, Reds, YlFn, YlFnBu, YlOrBr, YlOrRd, BrBg, PiYG, PRGn, PuOr, RdBu, RdGy, RdYlBu, RdYlGn, Spectral, Accent, Dark2, Paired, Pastel1, Pastel2, Set2, Set3)}
}
\value{
length(n) character vector of colors
}
\description{
Makes a lot of brewer colors using an "inexhaustible" brewer palette ie will not complain if number of colors requested is too high.

Yes - this technically violates the "grammar of graphics", but meant for quick and dirty use.
}
\author{
Marcin Imielinski
}
