% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R, R/gGraph.R
\name{c.junctions}
\alias{c.junctions}
\title{c -- concate junctions objects}
\usage{
\method{c}{junctions}(...)

\method{c}{junctions}(...)
}
\description{
c -- concate junctions objects

c -- concate junctions objects
}

