% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{c}
\alias{c}
\alias{c.gNode}
\alias{c.gEdge}
\alias{c.Junction}
\alias{c.gWalk}
\title{c.gNode}
\usage{
\method{c}{gNode}(...)

\method{c}{gEdge}(...)

\method{c}{Junction}(...)

\method{c}{gWalk}(..., force = FALSE)
}
\arguments{
\item{gNode}{objects}

\item{gEdge}{objects}

\item{Junction}{object}

\item{gWalk}{objects}
}
\value{
a new concatenated gNode

a new concatenated gEdge Object

a new concatenated Junction Object

a new concatenated gWalk Object
}
\description{
Concatenates gNode objects by node.id's stored in them. All arguments
must point at the same graph or an error will be thrown.

Concatenates gEdge by edge.id's stored in them. All arguments
must point at the same graph or an error will be thrown.

Concatenates Junction Objects

Concatenates gWalk objects by id's
}
\author{
Joe DeRose

Joe DeRose

Rick Mortensen
}
