% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{capitalize}
\alias{capitalize}
\title{capitalize}
\usage{
capitalize(string, un = FALSE)
}
\arguments{
\item{string}{character vector to capitalize}

\item{un}{logical flag whether to uncapitalize (=FALSE)}
}
\value{
character vector of strings with capitalized values
}
\description{
Capitalize first letter of each character element of vector "string"
}
