% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{cbind}
\alias{cbind}
\title{cbind wrapper}
\usage{
cbind(..., deparse.level = 1)
}
\arguments{
\item{...}{arguments to cbind}
}
\value{
vector of combined arguments
}
\description{
Forcing correct call of cbind
}
