% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{chr2num}
\alias{chr2num}
\title{chr2num}
\usage{
chr2num(x, xy = FALSE)
}
\arguments{
\item{x}{factor, Rle or character vector with chromosome names}

\item{xy}{Flag to convert M to 25, Y to 24 and X to 23. Default FALSE}
}
\value{
character vector with xy=FALSE, or numeric vector with xy=TRUE
}
\description{
Convert from chrXX to numeric format
}

