% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventCallers.R
\name{chromoplexy}
\alias{chromoplexy}
\title{Finds chromoplexy chains as clusters of "long distance" junctions that each span at least min.span
(i.e. distant regions on the reference) have junctions nearby ie within max.dist.
We filter to chains that have at least min.num footprints on the genome and involve at
least min.num long distance junctions.  We keep track of how many "other" (non small dup
and non small del) junctions there are in the vicinity for downstream filtering.}
\usage{
chromoplexy(
  gg,
  min.span = 1e+07,
  max.dist = 10000,
  min.num = 3,
  max.cn = 3,
  footprint.width = 1e+06,
  ignore.small.dups = TRUE,
  ignore.small.dels = TRUE,
  max.small = 50000,
  mark = FALSE,
  mark.col = "purple"
)
}
\arguments{
\item{gg}{gGraph}

\item{min.span}{minimimum span to define a "long distance" junction and also the span by which major footprints of the event must be separated}

\item{min.num}{minimum number of junctions and major footprints that define a chromoplexy}

\item{footprint.width}{padding around which to define the footprint of an event, note that the outputted footprint only includes the chromoplexy junction breakpoints}

\item{ignore.small.dups}{logical flag (FALSE) determining whether we ignore small dups when filtering on min.cushion}

\item{ignore.small.dels}{logical flag (FALSE) determining whether we ignore small dels when filtering on min.cushion}

\item{max.small}{threshold for calling a local dup or del "small" 1e4}
}
\value{
gGraph with $meta$chromoplexy annotated with chromoplexy event metadata and edges labeled with $chromoplexy id or NA if the edge does not belong to a chromoplexy
}
\description{
Finds chromoplexy chains as clusters of "long distance" junctions that each span at least min.span
(i.e. distant regions on the reference) have junctions nearby ie within max.dist.
We filter to chains that have at least min.num footprints on the genome and involve at
least min.num long distance junctions.  We keep track of how many "other" (non small dup
and non small del) junctions there are in the vicinity for downstream filtering.
}
