% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventCallers.R
\name{chromothripsis}
\alias{chromothripsis}
\title{Finds chromothripsis as clusters of >= min.seg segments and >= min.jun
junctions, clusters defined as clusters of segs with  <= max.seg.width
spread across <= max.major footprints with cn <= max.cn and
cn amplitude cn <= max.cn.amplitude 
and also fractional width distribution consistent with dirichlet 
orientation distribution consistent with Uniform and
window distribution (if multiple windows) consistent with Uniform.

"Major" footprints are defined as reduced footprints that
have width >= min.major.width.

We may also allow a number of (<= min.majo) "minor" footprints
(e.g. templated insertions) each <= min.major.width to prevent
clusters from being thrown out because certain TI junctions are mapped
in multiple parts.}
\usage{
chromothripsis(
  gg,
  min.seg = 8,
  min.jun = 7,
  max.cn = 4,
  max.cn.amplitude = 3,
  max.major = 4,
  min.major.width = 1e+05,
  max.minor = 2,
  min.width.p = 0.05,
  min.p.orientation = 0.001,
  min.stack = 3,
  min.mean.stack = 3,
  fbi.thresh = 50000,
  remove.small.junctions = TRUE,
  small.junction.thresh = 10000,
  scale.to.ploidy = TRUE,
  mark = TRUE,
  mark.col = "purple"
)
}
\arguments{
\item{gg}{gGraph}

\item{min.seg}{minimum number of segments in a CT (8)}

\item{min.jun}{minimum number of juctnions in a CT (8)}

\item{max.cn}{max CN of a CT (4)}

\item{max.major}{max number of "major" footprints of CT with width >= min.major.width}

\item{min.major.width}{width threshold defining a major footprint}

\item{max.minor}{max number of "minor" footprints of CT with width < min.mmajor.width}

\item{min.mean.stack}{average number of "stacks" in event treating each junction span as a GRanges (2)}

\item{mark}{logical flag whether to mark graph with CT events (FALSE)}

\item{mark.col}{logical flag of what to color events}

\item{width.thresh}{max width of a CT segment (in simplified gGraph) to consider (1e7)}

\item{max.cn.ampltitude}{max difference between top and bottom CN in a CT (3)}

\item{max.win}{max}
}
\value{
gGraph with nodes and edges annotated with integer chromothripsis event or NA and metadata showing some statistics for the returns chromothripsis events
}
\description{
Finds chromothripsis as clusters of >= min.seg segments and >= min.jun
junctions, clusters defined as clusters of segs with  <= max.seg.width
spread across <= max.major footprints with cn <= max.cn and
cn amplitude cn <= max.cn.amplitude 
and also fractional width distribution consistent with dirichlet 
orientation distribution consistent with Uniform and
window distribution (if multiple windows) consistent with Uniform.

"Major" footprints are defined as reduced footprints that
have width >= min.major.width.

We may also allow a number of (<= min.majo) "minor" footprints
(e.g. templated insertions) each <= min.major.width to prevent
clusters from being thrown out because certain TI junctions are mapped
in multiple parts.
}
