% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R, R/misc_utils.R
\name{colnames2}
\alias{colnames2}
\title{robust colnames

gives back character vector same number of columns of input regardless whether named or not}
\usage{
colnames2(x)

colnames2(x)
}
\arguments{
\item{str}{a path string}
}
\value{
a string with multiple parentheses replaced with a single parenthesis

a string with multiple parentheses replaced with a single parenthesis
}
\description{
robust colnames

gives back character vector same number of columns of input regardless whether named or not

robust colnames

gives back character vector same number of columns of input regardless whether named or not
}
\author{
Kevin Hadi
}
