% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{convertEdges}
\alias{convertEdges}
\title{convertEdges}
\usage{
convertEdges(nodes, edges, metacols = FALSE, cleanup = TRUE)
}
\arguments{
\item{nodes}{GRanges of signed nodes (ie $pnodes in gGraph object)}

\item{edges}{data.table of signed edges (ie $pedges in gGraph object)}
}
\description{
Takes nodes and edges and converts the edge table for the nodes if they were strandless
gEdge table will be the appropriate table for if we did:
     nodes[which(as.logical(strand(nodes)=="+"))]
     strand(nodes) = "*"
pre: Nodes have snode.id and index column (indicates index)
     edges have sedge.id
}
\author{
Joe DeRose
}
\keyword{internal}
