% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{copy}
\alias{copy}
\title{make deep copy, recursively

useful for dev
makes deep copy of R6 object, S4 object, or anything else really}
\usage{
copy(x, recurse_list = TRUE, depth = 0L)
}
\description{
make deep copy, recursively

useful for dev
makes deep copy of R6 object, S4 object, or anything else really
}
