% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R, R/misc_utils.R
\name{copy3}
\alias{copy3}
\title{make deep copy, recursively

useful for dev
makes deep copy of R6 object, S4 object, or anything else really}
\usage{
copy3(x, recurse_list = TRUE)

copy3(x, recurse_list = TRUE)
}
\description{
make deep copy, recursively

useful for dev
makes deep copy of R6 object, S4 object, or anything else really

make deep copy, recursively

useful for dev
makes deep copy of R6 object, S4 object, or anything else really
}
