% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGVdb.R
\name{createPGV}
\alias{createPGV}
\title{Creates a PGV instance}
\usage{
createPGV(directory = ".", inputgraphs = NA, cores = 10, ref = "hg19")
}
\description{
inside we have the add function to initialize the PGV datafile.json and get pointers to where all the files are
 also set up datafiles.json and get pointers to where the data folder and datafiles.json are being stored
 will need to set up datafiles.json and data folder along with backups for each
probably would get bloat backing up data folder so that is a bit suspect.
}
