% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R, R/misc_utils.R
\name{dedup.cols}
\alias{dedup.cols}
\title{applies dedup to colnames

dedup the column names of a data.frame/data.table}
\usage{
dedup.cols(tbl, remove = FALSE)

dedup.cols(tbl, remove = FALSE)
}
\value{
A data.table or data.frame

A data.table or data.frame
}
\description{
applies dedup to colnames

dedup the column names of a data.frame/data.table

applies dedup to colnames

dedup the column names of a data.frame/data.table
}
