% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{dedup}
\alias{dedup}
\title{dedup}
\usage{
dedup(x, suffix = ".")
}
\arguments{
\item{x}{input vector to dedup}

\item{suffix}{suffix separator to use before adding integer for dups in x}
}
\value{
length(x) vector of input + suffix separator + integer for dups and no suffix for "originals"
}
\description{
relabels duplicates in a character vector with .1, .2, .3
(where "." can be replaced by any user specified suffix)
}
\author{
Marcin Imielinski
}

