% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DIM}
\alias{DIM}
\title{Always returns a length 2 vector describing dimensions of input.}
\usage{
DIM(x)
}
\description{
NROW returns either length(x) or nrows(x) if x has dimensions.
NCOL2 returns 0 if length(x)
}
