% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{do.assign}
\alias{do.assign}
\title{Assign columns or list elements}
\usage{
do.assign(x, ..., pf = parent.frame())
}
\description{
Wrapper to assign list elements to tabular object.
Intended to be robust to S4 object specs.
}
\author{
Kevin Hadi
}
