% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{.cplex_customparams}
\alias{.cplex_customparams}
\title{param_lines = c(param_lines, paste("CPX_PARAM_WORKDIR", getwd(), sep = '\t'))}
\usage{
.cplex_customparams(out.file, numthreads = 0, nodefileind = NA,
  treememlim = NA)
}
\description{
param_lines = c(param_lines, paste("CPX_PARAM_WORKDIR", getwd(), sep = '\t'))
}
