% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{.e2class}
\alias{.e2class}
\title{.e2class}
\usage{
.e2class(K, eclass)
}
\description{
edge to contig class conversion
}
\details{
given matrix K of k contigs over e edges, each belonging to cardinality 1 or cardinality 2 equivalence classes,
assigns id's to equivalent contigs
}
