% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R, R/misc_utils.R
\name{dt_na2false}
\alias{dt_na2false}
\title{convert columns with NA to false

coerce NA in columns of class "logical" to FALSE}
\usage{
dt_na2false(dt, these_cols = NULL)

dt_na2false(dt, these_cols = NULL)
}
\arguments{
\item{dt}{data.table}

\item{these_cols}{NULL by default, will select columns of class logical, otherwise will be specified}
}
\value{
A data.table

A data.table
}
\description{
convert columns with NA to false

coerce NA in columns of class "logical" to FALSE

convert columns with NA to false

coerce NA in columns of class "logical" to FALSE
}
\author{
Kevin Hadi

Kevin Hadi
}
