% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{duplicated.matrix}
\alias{duplicated.matrix}
\title{R-3.5.1 version of duplicated.matrix}
\usage{
duplicated.matrix(x, incomparables = FALSE, MARGIN = 1, fromLast = FALSE, ...)
}
\value{
a logical
}
\description{
R-3.6.1 base::duplicated.matrix returns an array rather than a logical, breaking convex.basis
}
