% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{enframe_list}
\alias{enframe_list}
\alias{isNA}
\title{data table-ize list elements and add name column}
\usage{
isNA(x, na.char = c("NA", "NULL", "na", "null"))
}
\arguments{
\item{object}{tabled object}

\item{nm}{names of the new columns}
}
\value{
colnamed object
}
\description{



}
\author{
Kevin Hadi

Kevin Hadi

Kevin Hadi
}
