% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{==.gGraph}
\alias{==.gGraph}
\title{equals.gGraph}
\usage{
\method{==}{gGraph}(x, y)
}
\arguments{
\item{x}{gGraph}

\item{y}{gGraph}
}
\value{
TRUE if objects are equivalent
}
\description{
Returns TRUE if two graphs are equivalent by value based on having equivalent nodes and edges
(note this is different - intentionally - from identical, which will determine whether
the two variables are identical by reference i.e. point to the exact same location in memory)
}
