% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{==.gWalk}
\alias{==.gWalk}
\title{equals.gWalk}
\usage{
\method{==}{gWalk}(x, y)
}
\arguments{
\item{x}{gWalk}

\item{y}{gWalk}
}
\value{
TRUE if objects are equivalent
}
\description{
Returns TRUE if two walksets are equivalent, up to their reverse complement. Ignores order of input (to be implemented later!)
Walks are said to be equivalent iff their graphs are equivalent and the hashes of their node.ids are equivalent
}
