% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{etype}
\alias{etype}
\title{infer edge type based on node coordinates and orientation}
\usage{
etype(segs, es, force = FALSE, both = FALSE)
}
\arguments{
\item{segs}{the GRanges of nodes}

\item{es}{the data.table of edges}

\item{force}{logical, whether to overwrite existing node or edge type}

\item{both}{logical, if TRUE, return a list of updated segs and es}
}
\value{
es with "type" column.
}

