% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventCallers.R
\name{events.to.gr}
\alias{events.to.gr}
\title{Events to GRanges}
\usage{
events.to.gr(gg)
}
\arguments{
\item{gg}{gGraph}
}
\value{
GRanges containing ranges of annotated events along with all metadata 
from gg$meta$events
}
\description{
Extract event annotation as a GRanges
}
\author{
Alon Shaiber
}
