% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{fitcn}
\alias{fitcn}
\title{fitcn}
\usage{
fitcn(
  gw,
  cn.field = "cn",
  trim = TRUE,
  weight = NULL,
  obs.mat = NULL,
  verbose = TRUE,
  min.alt = TRUE,
  edgeonly = FALSE,
  evolve = FALSE,
  n.sol = 2,
  return.gw = TRUE,
  sep = "_"
)
}
\arguments{
\item{gw}{input gWalks}

\item{cn.field}{character column names of each graph's CN data}
}
\description{
fitcn
}
\author{
Julie Behr, Xiaotong Yao
}
