% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fra.duplicated}
\alias{fra.duplicated}
\title{Fast RA duplicate find}
\usage{
fra.duplicated(grl, pad = 500, ignore.strand = FALSE, sort_junctions = TRUE)
}
\arguments{
\item{pad}{Padding}

\item{ignore.strand}{Ignore breakend strands}
}
\description{
Function that supersedes ra.duplicated
}
\details{
See fra.overlaps for details on
implementation of fra.overlaps.
find_duplicates allows only non-self
matches to be considered inside fra.overlaps.
}
\author{
Kevin Hadi
}
