% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventCallers.R
\name{fusions}
\alias{fusions}
\title{fusions}
\usage{
fusions(
  graph = NULL,
  gencode = NULL,
  genes = NULL,
  annotate.graph = TRUE,
  mc.cores = 1,
  verbose = FALSE
)
}
\arguments{
\item{graph}{input gGraph}

\item{gencode}{GFF containing gene boundaries and exons, in similar format to
https://www.gencodegenes.org/}

\item{annotate.graph}{Annotate the graph generated Default: True}

\item{mc.cores}{number of cores to run. Default: 1}

\item{verbose}{output verbose argument to function Default: False}

\item{query}{optional query limiting walks to specific regions of interest}

\item{prom.window}{window to use around each transcript to identify putative 
promoter if promoter is NULL}
}
\value{
gWalks of gene fusions annotated with frame and gene(s)
}
\description{
annotates all gene fusions in gGraph relative to cds definitions

cds = gencode cds GRanges gff3 / gtf or GRangesList the latter (converted via rtracklayer::import)
which has fields $exon_number

"gene_name" GRangesList meta data field is used in annotation and in not 
creating "splice" fusionsthat arise from different transcripts of the same gene.
}
