% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{gEdge.in}
\alias{gEdge.in}
\title{subset gEdge on overlaps}
\usage{
\S4method{\%&\%}{gEdge}(x, y)
}
\arguments{
\item{x}{a gEdge}
}
\value{
the subset of gEdges overlapping with the query
}
\description{

}
