% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{gEdge.over}
\alias{gEdge.over}
\title{gEdge over}
\usage{
\S4method{\%^\%}{gEdge}(x, y)
}
\arguments{
\item{x}{a gEdge}
}
\value{
logical vector of length(x) which is TRUE if an overlap exists
}
\description{

}
