% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{gEdge}
\alias{gEdge}
\title{gEdge}
\description{
gEdge obejct
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-gEdge-new}{\code{gEdge$new()}}
\item \href{#method-gEdge-mark}{\code{gEdge$mark()}}
\item \href{#method-gEdge-subset}{\code{gEdge$subset()}}
\item \href{#method-gEdge-print}{\code{gEdge$print()}}
\item \href{#method-gEdge-clone}{\code{gEdge$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gEdge-new"></a>}}
\if{latex}{\out{\hypertarget{method-gEdge-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gEdge$new(seid = NULL, graph)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gEdge-mark"></a>}}
\if{latex}{\out{\hypertarget{method-gEdge-mark}{}}}
\subsection{Method \code{mark()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gEdge$mark(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gEdge-subset"></a>}}
\if{latex}{\out{\hypertarget{method-gEdge-subset}{}}}
\subsection{Method \code{subset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gEdge$subset(i)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gEdge-print"></a>}}
\if{latex}{\out{\hypertarget{method-gEdge-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gEdge$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gEdge-clone"></a>}}
\if{latex}{\out{\hypertarget{method-gEdge-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gEdge$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
