% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{.gencode_transcript_split}
\alias{.gencode_transcript_split}
\title{.gencode_transcript_split}
\usage{
.gencode_transcript_split(gr, tx)
}
\description{
splits gencode gr into transcript, taking care of some junky issues in the meantime
}
\author{
Marcin Imielinski
}
