% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{get.ploidy}
\alias{get.ploidy}
\title{get.ploidy
We define ploidy as the width-weighted mean of copy number. In other words, how many copies
of a unique set of genomic ranges are in the input.}
\usage{
get.ploidy(segs)
}
\arguments{
\item{segs}{a \code{GRanges} object holding \code{gGraph} node data with copy number annotated}
}
\value{
\code{numeric} scalar of the ploidy value
}
\description{
get.ploidy
We define ploidy as the width-weighted mean of copy number. In other words, how many copies
of a unique set of genomic ranges are in the input.
}

