% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_reciprocal_pairs}
\alias{get_reciprocal_pairs}
\title{get_reciprocal_pairs}
\usage{
get_reciprocal_pairs(jun, max_dist = 1000, distance_pad = 1e+05)
}
\arguments{
\item{max_dist}{Maximum distance between breakends of reciprocal pairs}

\item{distance_pad}{Max distance of query (breakends larger than this are not considered, limits compute)}
}
\description{
Get all reciprocal pairs from set of junctions
}
\details{
Match junction within 1e5
}
