% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome-package.R
\docType{package}
\name{gGnome-package}
\alias{gGnome-package}
\title{gGnome: An R API to manipulate genome graphs}
\description{
The gGnome package provides a flexible, queriable R interface to graphs and walks of reference genomic intervals. gGnome is written in the R6 object oriented standard and built around a powerful GenomicRanges, data.table, and igraph backend, and thus supports agile interaction with graphs consisting of hundreds of thousands of nodes and edges.
}
\author{
\strong{Maintainer}: Xiaotong Yao \email{xiaotong.yao23@gmail.com}

Authors:
\itemize{
  \item Marcin Imielinski \email{mai9037@med.cornell.edu} [contributor]
}

}
