% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\docType{class}
\name{gGraph-class}
\alias{gGraph-class}
\title{genomic rearrangement graph}
\usage{
gGraph$new(tile=NULL, junctions=NULL, cn = FALSE,
           jabba=NULL,
           weaver=NULL,
           prego=NULL,
           segs=NULL, es=NULL,
           ploidy=NULL, purity=NULL)

gg$nullGGraph()

gg$dipGraph(genome = NULL, chr = FALSE, include.junk = FALSE)

gg$karyograph(tile = NULL, juncs = NULL)

gg$addJuncs(junc)

gg$addSegs(bps)

gg$jab2gg(jabba, regular=NULL)

gg$
}
\arguments{
\item{tile}{the \code{GRanges} genome segmentation}

\item{junctions}{the \code{GRangesList} of SV junctions}

\item{cn}{\code{logical}}

\item{jabba}{the path to or actual \code{list} of \href{http://github.com/mskilab/JaBbA}{JaBbA} output}

\item{regular}{\code{logical}}
}
\details{
gGraph-class
}
