% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{gGraph.subset}
\alias{gGraph.subset}
\title{subset gGraph on overlaps}
\usage{
\S4method{\%&\%}{gGraph}(x, y)
}
\arguments{
\item{x}{a gGraph}

\item{y}{a GRanges or string coercible to one}
}
\value{
the portion of the gGraph that overlaps y (with loose ends wherever nodes / edges are broken)
}
\description{

}
