% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R, R/gGraph.R
\docType{data}
\name{gGraphJb}
\alias{gGraphJb}
\title{Descendant of gGraph class, where junction balance restraint must be met all the time}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
gGraphJb

gGraphJb
}
\description{
Descendant of gGraph class, where junction balance restraint must be met all the time

Descendant of gGraph class, where junction balance restraint must be met all the time
}
\keyword{datasets}

