% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{gNode over}
\alias{gNode over}
\title{subset gNode on overlaps}
\usage{
\S4method{\%^\%}{gNode}(x, y)
}
\arguments{
\item{x}{a gNode}
}
\value{
logical vector of length(x) specifying overlap
}
\description{

}
