% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{gNode subset}
\alias{gNode subset}
\title{subset gNode on overlaps}
\usage{
\S4method{\%&\%}{gNode}(x, y)
}
\arguments{
\item{x}{a gNode}
}
\value{
the subset of gNodes overlapping with the query
}
\description{
subset gNode on overlaps
}
