% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gNode.loose}
\alias{gNode.loose}
\title{gNode.loose}
\usage{
gNode.loose(gn, orientation)
}
\arguments{
\item{orientation}{(character) either '', 'right' or 'left'. By default returns all loose ends (orientation = ''). If 'right' or 'left' returns only the loose ends that are to the right or left of nodes, accordingly.}
}
\description{
internal

Internal utility function to get a GRanges with the right or left loose ends.
This function serves as the backend of gNode$loose
}
\author{
Alon Shaiber
}
