% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{gNode}
\alias{gNode}
\title{gNode}
\description{
gNode object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{gNode$new()}}
\item \href{#method-mark}{\code{gNode$mark()}}
\item \href{#method-rep}{\code{gNode$rep()}}
\item \href{#method-swap}{\code{gNode$swap()}}
\item \href{#method-subset}{\code{gNode$subset()}}
\item \href{#method-eval}{\code{gNode$eval()}}
\item \href{#method-clusters}{\code{gNode$clusters()}}
\item \href{#method-ego}{\code{gNode$ego()}}
\item \href{#method-print}{\code{gNode$print()}}
\item \href{#method-clone}{\code{gNode$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gNode$new(snode.id = NULL, graph)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mark"></a>}}
\if{latex}{\out{\hypertarget{method-mark}{}}}
\subsection{Method \code{mark()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gNode$mark(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rep"></a>}}
\if{latex}{\out{\hypertarget{method-rep}{}}}
\subsection{Method \code{rep()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gNode$rep(times)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-swap"></a>}}
\if{latex}{\out{\hypertarget{method-swap}{}}}
\subsection{Method \code{swap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gNode$swap(replacement)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subset"></a>}}
\if{latex}{\out{\hypertarget{method-subset}{}}}
\subsection{Method \code{subset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gNode$subset(i)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eval"></a>}}
\if{latex}{\out{\hypertarget{method-eval}{}}}
\subsection{Method \code{eval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gNode$eval(x, right = TRUE, all = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clusters"></a>}}
\if{latex}{\out{\hypertarget{method-clusters}{}}}
\subsection{Method \code{clusters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gNode$clusters(mode = "weak")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ego"></a>}}
\if{latex}{\out{\hypertarget{method-ego}{}}}
\subsection{Method \code{ego()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gNode$ego(order = 0, mindist = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gNode$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
