% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr.breaks.R
\name{gr.breaks}
\alias{gr.breaks}
\title{Break GRanges at given breakpoints into disjoint gr}
\usage{
gr.breaks(query, bps = NULL)
}
\arguments{
\item{query}{a disjoint \code{GRanges} object to be broken}

\item{bps}{\code{GRanges} of width 1, locations of the bp}
}
\value{
\code{GRanges} disjoint object at least the same length as query,
with a metadata column \code{qid} indicating input index where new segment is from
}
\description{
Break GRanges at given breakpoints into disjoint gr
}
\examples{


}
