% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R, R/misc_utils.R
\name{gr.noval}
\alias{gr.noval}
\title{get rid of mcols on GRanges/GRangesLists}
\usage{
gr.noval(gr, keep.col = NULL, drop.col = NULL)

gr.noval(gr, keep.col = NULL, drop.col = NULL)
}
\value{
GRanges or GRangesList

GRanges or GRangesList
}
\description{
remove all metadata from GRanges or GRangesList

remove all metadata from GRanges or GRangesList
}
\author{
Kevin Hadi

Kevin Hadi
}
