% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gr_deconstruct_by}
\alias{gr_deconstruct_by}
\title{Removing by field and random string barcode to seqnames for more efficient by queries}
\usage{
gr_deconstruct_by(x, by = NULL, meta = FALSE)
}
\value{
A GRanges with the by metadata field attached to the seqnames
}
\description{
To be used with gr_construct_by
}
\author{
Kevin Hadi
}
