% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{\%&\%}
\alias{\%&\%}
\alias{\%&\%,}
\alias{gEdge}
\alias{Method}
\title{subset x on y edges}
\usage{
\S4method{\%&\%}{gNode}(x, y)
}
\value{
subset of gEdge that overlaps
}
\description{
gEdge1 %&% gEdge2/Junction returns the subsets of gEdge1 that overlaps gEdge2/Junction
}
\author{
Rick Mortenson
}
