% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{\%K\%}
\alias{\%K\%}
\title{similar to setkey except a general use utility

slower version of setkey, but for interactive use}
\usage{
thisx \%K\% thisy
}
\description{
similar to setkey except a general use utility

slower version of setkey, but for interactive use
}
\author{
Kevin Hadi
}
