% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{grl.match}
\alias{grl.match}
\title{grl.match
Matching the GRanges elements in a GRangesList}
\usage{
grl.match(grl1, grl2, ordered = FALSE, ignore.strand = FALSE,
  verbose = FALSE)
}
\arguments{
\item{ordered}{if TRUE the order of elements are considered in comparison}

\item{ignore.strand}{if TRUE the strand info is ignored}
}
\value{
vector same length as grl1, containing the indices of the first matching object in grl2
}
\description{
grl.match
Matching the GRanges elements in a GRangesList
}
