% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{gtf2json}
\alias{gtf2json}
\title{gtf2json
Turning a GTF format gene annotation into JSON}
\usage{
gtf2json(gtf = NULL, gtf.rds = NULL, gtf.gr.rds = NULL,
  filename = "./gtf.json", genes = NULL, grep = NULL, grepe = NULL,
  chrom.sizes = NULL, include.chr = NULL, gene.collapse = TRUE,
  verbose = TRUE)
}
\description{
gtf2json
Turning a GTF format gene annotation into JSON
}
