% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{gWalk.over}
\alias{gWalk.over}
\title{subset gWalk on overlaps}
\usage{
\S4method{\%^\%}{gWalk}(x, y)
}
\arguments{
\item{x}{a gWalk}
}
\value{
logical vector of length(x) which is TRUE if an overlap exists
}
\description{
subset gWalk on overlaps
}
