% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{gWalk}
\alias{gWalk}
\title{gWalk}
\description{
gWalk object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-gWalk-new}{\code{gWalk$new()}}
\item \href{#method-gWalk-set}{\code{gWalk$set()}}
\item \href{#method-gWalk-subset}{\code{gWalk$subset()}}
\item \href{#method-gWalk-dts}{\code{gWalk$dts()}}
\item \href{#method-gWalk-fix}{\code{gWalk$fix()}}
\item \href{#method-gWalk-rep}{\code{gWalk$rep()}}
\item \href{#method-gWalk-print}{\code{gWalk$print()}}
\item \href{#method-gWalk-json}{\code{gWalk$json()}}
\item \href{#method-gWalk-eval}{\code{gWalk$eval()}}
\item \href{#method-gWalk-disjoin}{\code{gWalk$disjoin()}}
\item \href{#method-gWalk-simplify}{\code{gWalk$simplify()}}
\item \href{#method-gWalk-mark}{\code{gWalk$mark()}}
\item \href{#method-gWalk-fitcn}{\code{gWalk$fitcn()}}
\item \href{#method-gWalk-gtrack}{\code{gWalk$gtrack()}}
\item \href{#method-gWalk-clone}{\code{gWalk$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gWalk-new"></a>}}
\if{latex}{\out{\hypertarget{method-gWalk-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gWalk$new(
  snode.id = NULL,
  sedge.id = NULL,
  grl = NULL,
  graph = NULL,
  meta = NULL,
  circular = NULL,
  disjoin = FALSE,
  drop = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gWalk-set"></a>}}
\if{latex}{\out{\hypertarget{method-gWalk-set}{}}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gWalk$set(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gWalk-subset"></a>}}
\if{latex}{\out{\hypertarget{method-gWalk-subset}{}}}
\subsection{Method \code{subset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gWalk$subset(i)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gWalk-dts"></a>}}
\if{latex}{\out{\hypertarget{method-gWalk-dts}{}}}
\subsection{Method \code{dts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gWalk$dts(ix = 1:self$length, makelists = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gWalk-fix"></a>}}
\if{latex}{\out{\hypertarget{method-gWalk-fix}{}}}
\subsection{Method \code{fix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gWalk$fix(pattern = NULL, replacement = NULL, drop = TRUE, seqlengths = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gWalk-rep"></a>}}
\if{latex}{\out{\hypertarget{method-gWalk-rep}{}}}
\subsection{Method \code{rep()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gWalk$rep(times)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gWalk-print"></a>}}
\if{latex}{\out{\hypertarget{method-gWalk-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gWalk$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gWalk-json"></a>}}
\if{latex}{\out{\hypertarget{method-gWalk-json}{}}}
\subsection{Method \code{json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gWalk$json(
  filename = ".",
  save = TRUE,
  verbose = FALSE,
  annotations = NULL,
  nfields = NULL,
  efields = NULL,
  stack.gap = 1e+05,
  include.graph = TRUE,
  settings = list(y_axis = list(title = "copy number", visible = TRUE)),
  cid.field = NULL,
  no.y = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gWalk-eval"></a>}}
\if{latex}{\out{\hypertarget{method-gWalk-eval}{}}}
\subsection{Method \code{eval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gWalk$eval(x, node, edge)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gWalk-disjoin"></a>}}
\if{latex}{\out{\hypertarget{method-gWalk-disjoin}{}}}
\subsection{Method \code{disjoin()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gWalk$disjoin(
  gr = NULL,
  graph = NULL,
  by = NULL,
  collapse = TRUE,
  na.rm = TRUE,
  avg = FALSE,
  sep = ",",
  FUN = default.agg.fun.generator(na.rm = na.rm, sep = sep, avg = avg)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gWalk-simplify"></a>}}
\if{latex}{\out{\hypertarget{method-gWalk-simplify}{}}}
\subsection{Method \code{simplify()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gWalk$simplify(
  by = NULL,
  na.rm = TRUE,
  avg = FALSE,
  sep = ",",
  FUN = default.agg.fun.generator(na.rm = na.rm, sep = sep, avg = avg)
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gWalk-mark"></a>}}
\if{latex}{\out{\hypertarget{method-gWalk-mark}{}}}
\subsection{Method \code{mark()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gWalk$mark(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gWalk-fitcn"></a>}}
\if{latex}{\out{\hypertarget{method-gWalk-fitcn}{}}}
\subsection{Method \code{fitcn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gWalk$fitcn(
  trim = TRUE,
  weight = NULL,
  obs.mat = NULL,
  verbose = FALSE,
  min.alt = TRUE,
  edgeonly = FALSE,
  evolve = FALSE,
  n.sol = 2
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gWalk-gtrack"></a>}}
\if{latex}{\out{\hypertarget{method-gWalk-gtrack}{}}}
\subsection{Method \code{gtrack()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gWalk$gtrack(name = NULL, stack.gap = 1e+05, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gWalk-clone"></a>}}
\if{latex}{\out{\hypertarget{method-gWalk-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gWalk$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
