% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\docType{data}
\name{gWalks}
\alias{gWalks}
\title{gWalks: subclass to gGraph}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
gWalks
}
\description{
gWalks: subclass to gGraph
}
\keyword{datasets}
