% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{haplograph}
\alias{haplograph}
\title{haplograph}
\usage{
haplograph(walks, breaks = NULL)
}
\arguments{
\item{walks}{gWalk or GRangesList (with seqinfo fully populated)}
}
\description{
Haplograph creates a graph from a set of walks each which is joined
to a reference graph backbone
i.e. each haplotype is a bubble on the original reference graph

Haplotype ends also branch to each other, as long as the termini
of both haplotype's end nodes are contained in the other end node.
}
