% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\docType{data}
\name{hGraph}
\alias{hGraph}
\title{hGraph: haplotype gGraph (underdeveloped)
has to be balanced and one unique path through all segs}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
hGraph
}
\description{
hGraph: haplotype gGraph (underdeveloped)
has to be balanced and one unique path through all segs
}
\keyword{datasets}

