% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{intersect.gEdge}
\alias{intersect.gEdge}
\title{intersect.gEdge}
\usage{
intersect.gEdge(x, y)
}
\arguments{
\item{x}{a gEdge Object}

\item{y}{a gEdge Object}
}
\value{
new gEdge containing the intersection of x and y
}
\description{
Returns a new gEdge which is the intersection of x and y (edge.id's).
All arguments must point at the same graph or an error will be thrown.
}
\author{
Joe DeRose
}
