% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{intersect.gNode}
\alias{intersect.gNode}
\title{intersect.gNode}
\usage{
intersect.gNode(x, y)
}
\arguments{
\item{x}{a gNode Object}

\item{y}{a gNode Object}
}
\value{
new gNode Object containing the intersection of x and y
}
\description{
Returns a new gNode which is the intersection of x and y (node.id's).
All arguments must point at the same graph or an error will be thrown.
}
\author{
Joe DeRose
}
