% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{intersect.Junction}
\alias{intersect.Junction}
\title{intersect.Junction}
\usage{
intersect.Junction(x, y, pad = 0, ...)
}
\arguments{
\item{x}{a Junction Object}

\item{y}{a Junction Object}
}
\value{
new Junction Object containing the intersection of x and y
}
\description{
Returns a new Junction Object which is the intersection of x and y.
}
\author{
Rick Mortensen
}
