% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{isInteger}
\alias{isInteger}
\title{Testing if a numeric value is integer}
\usage{
isInteger(x, eps = 1e-300)
}
\arguments{
\item{x}{numeric vector}

\item{eps}{infinitely small positive number}
}
\value{
logical vector of same length
}
\description{
Testing if a numeric value is integer
}
\author{
Xiaotong Yao
}
