% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R, R/misc_utils.R
\name{isNA}
\alias{isNA}
\title{is.na but also tests for "NA" character}
\usage{
isNA(x, na.char = c("NA", "NULL", "na", "null"))

isNA(x, na.char = c("NA", "NULL", "na", "null"))
}
\description{



}
\author{
Kevin Hadi

Kevin Hadi
}
