% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{jab2json}
\alias{jab2json}
\title{jab2json}
\usage{
jab2json(jab, file, maxcn = 100, maxweight = 100)
}
\arguments{
\item{jab}{input jab object}

\item{file}{output json file}
}
\description{
Dumps JaBbA graph into json
}
\author{
Marcin Imielinski
}
