% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{jabba.gwalk}
\alias{jabba.gwalk}
\title{jabba.gwalk}
\usage{
jabba.gwalk(jab, verbose = FALSE, return.grl = TRUE)
}
\arguments{
\item{jab}{JaBbA object}
}
\value{
GRangesList of walks with copy number as field $cn, cyclic walks denoted
as field $is.cycle == TRUE, and $wid (width) and $len (segment length) of walks
as additional metadata
}
\description{
Computes greedy collection (i.e. assembly) of genome-wide walks (graphs and cycles) by finding shortest paths in JaBbA graph.
}
\author{
Marcin Imielinski

Xiaotong Yao
}

