% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{jbaLP}
\alias{jbaLP}
\title{jbaLP}
\usage{
jbaLP(
  kag.file = NULL,
  kag = NULL,
  cn.field = "cnmle",
  var.field = "raw.var",
  min.var = 0.001,
  lambda = 10,
  L0 = TRUE,
  loose.collapse = FALSE,
  M = 1000,
  verbose = 1,
  tilim = 1000,
  epgap = 0.001
)
}
\arguments{
\item{kag.file}{(character)}

\item{kag}{(karyograph object)}

\item{cn.field}{(character) column in karyograph with CN guess, default cnmle}

\item{var.field}{(character) column in karyograph with node weight guess, default}

\item{lambda}{(numeric) slack penalty, default 10}

\item{L0}{(logical) default TRUE}

\item{loose.collapse}{(logical) default FALSE}

\item{M}{(numeric) max CN}

\item{verbose}{(numeric) 0 (nothing) 1 (everything but MIP) 2 (print the MIP), default 1}

\item{tilim}{(numeric) default 1e3}

\item{epgap}{(numeric) default 1e-3}
}
\description{
jbaLP

Simple (probably temporary) wrapper around balance for JaBbA LP
Reads karyograph.rds file and balances it using cnmle as CN estimate
}
\author{
Marcin Imielinski, Zi-Ning Choo
}
