% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{jJ}
\alias{jJ}
\title{parse junctions from a variety of formats}
\usage{
jJ(rafile,
keep.features = T,
seqlengths = NULL,
chr.convert = T,
geno=NULL,
flipstrand = FALSE,
swap.header = NULL,
breakpointer = FALSE,
seqlevels = NULL,
force.bnd = FALSE,
skip = NA)
}
\arguments{
\item{rafile}{path to the junctions file. See details for the compatible formats.}

\item{keep.features}{\code{logical}, if TRUE preserve meta data from the input}

\item{seqlengths}{a named \code{numeric} vector containing reference contig lengths}

\item{chr.convert}{\code{logical}, if TRUE strip "chr" prefix from contig names}

\item{geno}{\code{logical}, whether to parse the 'geno' fields of VCF}

\item{hg}{\code{character}, human genome version}

\item{flipstrand}{\code{logical}, if TRUE will flip breakpoint strand}

\item{swap.header}{path to the alternative VCF header file}

\item{breakpointer}{\code{logical}, if TRUE will parse as breakpointer output}

\item{seqlevels}{vector for renaming the chromosomes}

\item{force.bnd}{if TRUE overwrite all junction "type" to "BND"}

\item{skip}{\code{numeric} lines to skip}
}
\value{
a \code{Junction} object
}
\description{
Parsing various formats of structural variation data into junctions.
}
\details{
A junction is a unordered pair of strand-specific genomic locations (breakpoints). Within a given
reference genome coordinate system, we call the direction in which coordinates increase "+". A breakpoint
is a width 1 (\code{start==end})genomic range with \code{strand} specified, and "+" means the side with larger
coordinate is fused with the other breakpoint in a junction.

\code{rafile} must be one of the following formats:
1) Some VCF (variant call format). We currently support the VCF output
from a number of structural variation detection methods, namely
SvABA (https://github.com/walaj/svaba),
DELLY (https://github.com/dellytools/delly),
LUMPY (https://github.com/arq5x/lumpy-sv),
novoBreak (https://sourceforge.net/projects/novobreak/). In theory,
VCF defined with BND style should be compatible but be cautious
when using the output from other methods since
no universal data definition is adopted by the community yet.
2) BEDPE (http://bedtools.readthedocs.io/en/latest/content/general-usage.html#bedpe-format)
3) Textual output from Breakpointer
(http://archive.broadinstitute.org/cancer/cga/breakpointer)
4) R serialized object storing junctions (.rds)
}
\section{Warning}{

We assume the orientation definition in the input is consistent with ours. Check with
the documentation of your respective method to make sure. If the contrary, use
\code{flipstrand=TRUE} to reconcile.
}

\author{
Xiaotong Yao
}
