% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{Junction class constructor}
\alias{Junction class constructor}
\alias{Junction}
\title{set}
\description{
set

set
}
\author{
Joe DeRose and Rick Mortensen

Rick Mortensen
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Junction-new}{\code{Junction$new()}}
\item \href{#method-Junction-removeDups}{\code{Junction$removeDups()}}
\item \href{#method-Junction-gw}{\code{Junction$gw()}}
\item \href{#method-Junction-subset}{\code{Junction$subset()}}
\item \href{#method-Junction-set}{\code{Junction$set()}}
\item \href{#method-Junction-print}{\code{Junction$print()}}
\item \href{#method-Junction-clone}{\code{Junction$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Junction-new"></a>}}
\if{latex}{\out{\hypertarget{method-Junction-new}{}}}
\subsection{Method \code{new()}}{
Builds junction class, grl must be GRangesList with each GRanges of length 2 JUNCTION
Empty junctions are removed
If grl is empty GRangesList, Junctions class is empty
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Junction$new(grl = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{grl}}{GRangesList of signed breakpoint pairs}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Junction-removeDups"></a>}}
\if{latex}{\out{\hypertarget{method-Junction-removeDups}{}}}
\subsection{Method \code{removeDups()}}{
Removes duplicate junctions in this Junction Object

this Junction Object has no duplicate junctions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Junction$removeDups()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Junction-gw"></a>}}
\if{latex}{\out{\hypertarget{method-Junction-gw}{}}}
\subsection{Method \code{gw()}}{
Returns a gWalk representing the left -> right traversal of this junction +- padding
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Junction$gw(pad = 0)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
GRangesList of the junctions in this Junction Object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Junction-subset"></a>}}
\if{latex}{\out{\hypertarget{method-Junction-subset}{}}}
\subsection{Method \code{subset()}}{
Allows subseting of the Junction object using bracket notation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Junction$subset(i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{integer or self$length logical vector specifying subset}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Junction-set"></a>}}
\if{latex}{\out{\hypertarget{method-Junction-set}{}}}
\subsection{Method \code{set()}}{
set metadata for junction object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Junction$set(...)}\if{html}{\out{</div>}}
}

\subsection{Details}{
sets metadata to either a scalar or vector
where the vector is the same length as the junction object
```
jj$set(cn = 7)
jj$set(cn = c(1,2,3))
jj$set(col = c("red", "blue", "green"))
```
(after setting, metadata is accessible through $dt accessor)
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Junction-print"></a>}}
\if{latex}{\out{\hypertarget{method-Junction-print}{}}}
\subsection{Method \code{print()}}{
Prints out the Junction Object. Prints the length and the GRangesList of the junctions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Junction$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Junction-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Junction-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Junction$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
