% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{Junction.over}
\alias{Junction.over}
\title{Junction over}
\usage{
\S4method{\%^\%}{Junction}(x, y)
}
\arguments{
\item{x}{a Junction}

\item{y}{a Junction, gEdge, GenomicRanges, string}
}
\value{
logical vector of length(x) which is TRUE if an overlap exists
}
\description{
Junction over
}
