% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{Junction.subset}
\alias{Junction.subset}
\title{subset gEdge on overlaps}
\usage{
\S4method{\%&\%}{Junction}(x, y)
}
\arguments{
\item{x}{a Junction}

\item{y}{a Junction, gEdge, GenomicRanges, string}
}
\value{
the subset of gEdges overlapping with the query
}
\description{
subset gEdge on overlaps
}
