% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\docType{class}
\name{junctions-class}
\alias{junctions-class}
\alias{junctions}
\title{TODO:}
\description{
simplify
decouple
add
subtract
gg$junctions, gg$ab.edges
junctions: ra.merge, ra.dedup, ra.dist, ra.equal
gwalks: as.gGraph, write.json
jgraph
find.fusion
proximity
}
\details{
Naming:
-- use S3 to overload exported function names with pure lowercase
-- the exposed fields should have lower case names too
-- Okay to keep the camel cases of object or internal method names for now,
to replace gradually later
-- arithmetics of graphs
-- ## TODO: develop the most efficient way to r/w GFA1 format

Structure:
-- use S4 to extend GRL class to junctions, don't use R6

Documentation:
-- define exported functions and for each come up with a short use case

Final destination: I want something like
-- jab = read.jab("jabba.simple.rds"); and jab is a bGraph unless told otherwise
-- gw = gwalk(jab); and gw is a gWalk unless told otherwise
-- plot(jab) plots the default gTrack and returns in situ
-- write.json(jab); write.json(gw); saves the JSON format for viz
-- hood(jab, win, d=1e6); returns the +/-1Mb neighborhood of the
-- subgraph(jab, expr); returns the subgraph where the nodes evaluate to TRUE in expr
--
junctions
S4 wrapper around GRangesList to store junction info
}
