% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{karyotype}
\alias{karyotype}
\title{karyotype}
\usage{
karyotype(karyo = NULL, cytoband = NULL, ...)
}
\arguments{
\item{karyo}{karyotype string to generate a gWalk of alleles representing karyotype}

\item{cytoband}{path or URL to UCSC style cytoband file}

\item{...}{Additional arguments sent to the \code{gTrack} constructor}
}
\description{
returns gWalk (if karyo arg is not NULL) or gGraph of
cytoBands given chrom.sizes file, with built in colormap
for disjoining with other gGraphs and visualizing cytobands
in the context of a gGRaph
}
\author{
Marcin Imielinski
}
