% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{label.runs}
\alias{label.runs}
\title{label.runs}
\usage{
label.runs(x)
}
\description{
For logical input labels all instances of "TRUE" with a unique label and everything else as false

For non-logical (e.g. character) input labels, labels each contiguous runs of the same value with a unique label
(note: even subsequent runs of an earlier used value in the vector will be given a new unique label)
}
\author{
Marcin Imielinski
}
