% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{levapply}
\alias{levapply}
\title{levapply}
\usage{
levapply(x, by, FUN = "order")
}
\arguments{
\item{x}{input vector of data}

\item{by}{length(x) vector of categorical labels}

\item{FUN}{function that takes a length k vector and outputs a length k vector, used for processing each "level" of by}
}
\value{
length(x) vector of outputs, the results of applying FUN to each "by" defined level of x
}
\description{
Applies FUN locally to levels of x and returns vector of length()
(eg can do a "local" order within levels)
}
\author{
Marcin Imielinski
}
