% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{loosefix}
\alias{loosefix}
\title{loosefix}
\usage{
loosefix(gg)
}
\arguments{
\item{gg}{gGraph with node and edge fields 'cn', node field loose.cn.left and loose.cn.right}
}
\description{
Updates loose.left, loose.right, loose.cn.left, loose.cn.right based on value of cn field
}
\author{
Marcin Imielinski
}
