% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{merge}
\alias{merge}
\title{merge for undefined number of Junction objects}
\usage{
merge(...)

merge(...)
}
\description{
merge for undefined number of Junction objects

merge for undefined number of Junction objects
}
