% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{merge.repl}
\alias{merge.repl}
\title{merging data tables with collapsing columns with the same name

Merge two data tables with various replacing strategies
for columns common between x and y that are not used to merge
(i.e. not specified in the "by" argument)}
\usage{
\method{merge}{repl}(
  dt.x,
  dt.y,
  sep = "_",
  replace_NA = TRUE,
  force_y = TRUE,
  overwrite_x = FALSE,
  keep_order = FALSE,
  keep_colorder = TRUE,
  keep_factor = TRUE,
  ...
)
}
\arguments{
\item{replace_NA}{logical, only use values in dt.y, any dt.x not in dt.y is clobbered (NA)}

\item{force_y}{logical, should x and y common columns be merged?}

\item{overwrite_x}{logical, if force_y = TRUE, should NA values in y replace x?}
}
\value{
A data.table
}
\description{
merging data tables with collapsing columns with the same name

Merge two data tables with various replacing strategies
for columns common between x and y that are not used to merge
(i.e. not specified in the "by" argument)
}
\author{
Kevin Hadi
}
