% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{mmatch}
\alias{mmatch}
\title{mmatch}
\usage{
mmatch(A, B, dir = 1)
}
\arguments{
\item{A}{query matrix k1 x n}

\item{B}{subject matrix k2 x n}

\item{dir}{1}
}
\value{
length k1 vector specifying first row of B matching row i of A
}
\description{
match rows of matrix A to matrix B
}
\author{
Marcin Imielinski
}
