% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MULTIDIM}
\alias{MULTIDIM}
\title{Returns a length 2 vector describing first two dimensions of input, and .}
\usage{
MULTIDIM(x)
}
\description{
NROW returns either length(x) or nrows(x) if x has dimensions.
NCOL2 returns 0 if length(x)
}
