% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{munlist}
\alias{munlist}
\title{munlist}
\usage{
munlist(x, force.rbind = F, force.cbind = F, force.list = F)
}
\arguments{
\item{x}{list of vectors, matrices, or data frames}

\item{force.rbind}{logical flag to force concatenation via rbind (=FALSE), otherwise will guess}

\item{force.cbind}{logical flag to force concatenation via cbind (=FALSE), otherwise will guess}

\item{force.list}{logical flag to force concatenation via unlist (=FALSE), otherwise will guess}
}
\value{
data.frame of concatenated input data with additional fields $ix and $iix specifying the list item and within-list index from which the given row originated from
}
\description{
unlists a list of vectors, matrices, data frames into a n x k matrix
whose first column specifies the list item index of the entry
and second column specifies the sublist item index of the entry
and the remaining columns specifies the value(s) of the vector
or matrices.

force.cbind = T will force concatenation via 'cbind'
force.rbind = T will force concatenation via 'rbind'
}
\author{
Marcin Imielinski
}
