% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R, R/misc_utils.R
\name{na2false}
\alias{na2false}
\title{replace logical vector with NA to FALSE}
\usage{
na2false(v)

na2false(v)
}
\value{
A logical vector with NAs set to FALSE

A logical vector with NAs set to FALSE
}
\description{
A convenience function to set a logical vector with NAs to false

A convenience function to set a logical vector with NAs to false
}
\author{
Kevin Hadi

Kevin Hadi
}
