% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{NCOL2}
\alias{NCOL2}
\title{Ensuring 2nd dimension of empty vector or NULL is 1}
\usage{
NCOL2(x)
}
\description{
NCOL = 1 for NULL, or any vector with length == 0.
Seems counterintuitive so this is the fix
}
\author{
Kevin Hadi
}
