% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{peel}
\alias{peel}
\title{peel}
\usage{
peel(gg, field = NULL, embed.loops = FALSE, verbose = FALSE)
}
\arguments{
\item{gg}{gGraph with field cn}

\item{field}{edge metadata field to use to rank walk solutions (default edge field type == 'ALT')s}

\item{embed.loops}{logical flag (FALSE) if TRUE will embed all the loops in the output into an arbitrary linear (path) walk}

\item{verbose}{flag = 1 regular verbosity, 2 = dump out Rcplex traces}
}
\value{
collection of gWalks annotated with cn on the original that when added will give the marginal copy profile on inputted nodes and edges
}
\description{
Greedily "peels" walks off a genome graph with node and edge field "cn"
by finding successive flows that maximizes some edge metadata field (specified as field)
which if not specified will be a logical field type == 'ALT' specifying whether that
junction is an ALT edge
}
\author{
Marcin Imielinski
}
