% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGVdb.R
\name{PGVdb}
\alias{PGVdb}
\title{work horse graph function}
\description{
work horse graph function

work horse graph function
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PGVdb-new}{\code{PGVdb$new()}}
\item \href{#method-PGVdb-addgraphs}{\code{PGVdb$addgraphs()}}
\item \href{#method-PGVdb-push}{\code{PGVdb$push()}}
\item \href{#method-PGVdb-refreshPGV}{\code{PGVdb$refreshPGV()}}
\item \href{#method-PGVdb-clone}{\code{PGVdb$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGVdb-new"></a>}}
\if{latex}{\out{\hypertarget{method-PGVdb-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGVdb$new(datafiles.json = NA, datafolder = NA, PGV_public_dir = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGVdb-addgraphs"></a>}}
\if{latex}{\out{\hypertarget{method-PGVdb-addgraphs}{}}}
\subsection{Method \code{addgraphs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGVdb$addgraphs(self, PGVInput, cores = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{PGVInput}}{PGVInput class}

\item{\code{cores}}{number of cores to use
push object}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
accepts PGVInput class and pushes it
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGVdb-push"></a>}}
\if{latex}{\out{\hypertarget{method-PGVdb-push}{}}}
\subsection{Method \code{push()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGVdb$push(self, backup = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Details}{
pushes object

Refresh object
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGVdb-refreshPGV"></a>}}
\if{latex}{\out{\hypertarget{method-PGVdb-refreshPGV}{}}}
\subsection{Method \code{refreshPGV()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGVdb$refreshPGV(self)}\if{html}{\out{</div>}}
}

\subsection{Details}{
refreshes object from source given self. Used for any
changes since the object was deserialized
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGVdb-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PGVdb-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGVdb$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
