% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGVdb.R
\name{PGVInput}
\alias{PGVInput}
\title{init: initializes table}
\description{
init: initializes table

init: initializes table
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PGVInput-new}{\code{PGVInput$new()}}
\item \href{#method-PGVInput-add}{\code{PGVInput$add()}}
\item \href{#method-PGVInput-clone}{\code{PGVInput$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGVInput-new"></a>}}
\if{latex}{\out{\hypertarget{method-PGVInput-new}{}}}
\subsection{Method \code{new()}}{
adds graphs to table
validate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGVInput$new(table = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGVInput-add"></a>}}
\if{latex}{\out{\hypertarget{method-PGVInput-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGVInput$add(new.graphs)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PGVInput-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PGVInput-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PGVInput$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
