% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{phased.binstats}
\alias{phased.binstats}
\title{phased.binstats}
\usage{
phased.binstats(
  gg,
  bins = NULL,
  purity = NULL,
  ploidy = NULL,
  count.field = "count",
  allele.field = "allele",
  phase.blocks = NULL,
  edge.phase.dt = NULL,
  vbase.count.thres = 5,
  vbase.prop.thres = 0.9,
  min.bins = 3,
  min.var = 0.001,
  verbose = TRUE,
  mc.cores = 8
)
}
\arguments{
\item{gg}{gGraph}

\item{bins}{GRanges with:}

\item{purity}{(numeric)}

\item{ploidy}{(numeric)}

\item{count.field}{(str) field containing allele read counts (default count)}

\item{allele.field}{(str) field for containing allele label for read counts (default allele)}

\item{phase.blocks}{(GRanges) GRanges containing phase blocks (e.g. from HAPCUT). default NULL.}

\item{edge.phase.dt}{(data.table) with columns n1.major, n2.major, n1.minor, n2.minor and edge.id providing major/minor allele counts}

\item{vbase.count.thres}{(int) number of variant base counts required to phase edges (default 5)}

\item{vbase.prop.thres}{(float) proportion of allele excess required to phase edges (default 0.9)}

\item{min.bins}{(numeric) minimum number of bins for intra segment variance (default 3)}

\item{min.var}{(numeric) min allowable variance (default 0.1)}

\item{verbose}{(bool) default TRUE for debugging}

\item{mc.cores}{(int) number of cores}
}
\value{
gGraph whose nodes are annotated with $cn.major, $cn.minor, $haplotype, and $weight fields
}
\description{
Given GRanges containing major/minor allele counts and a balanced but unphased gGraph,
prepares phased gGraph input to balance.
}
