% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{phased.postprocess}
\alias{phased.postprocess}
\title{phased.postprocess}
\usage{
phased.postprocess(gg, phase.blocks = NULL, mc.cores = 8, verbose = 1)
}
\arguments{
\item{gg}{junction-balanced phased gGraph. each node must have associated og node.id}

\item{phase.blocks}{(GRanges) granges of phase blocks from linked reads. default = NULL}

\item{mc.cores}{(int) number of cores. default = 8.}

\item{verbose}{(bool) verbose > 0 prints stuff. default 1.}
}
\description{
Postprocess junction-balanced phased graph and creates unphased regions
This identifies regions without allelic CN imbalance
}
