% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{ploidy}
\alias{ploidy}
\title{ploidy

Computes ploidy i.e. average CN for a gGraph}
\usage{
ploidy(gg)
}
\arguments{
\item{gg}{gGraph}
}
\description{
ploidy

Computes ploidy i.e. average CN for a gGraph
}
\author{
Marcin Imielinski
}
