% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{proximity}
\alias{proximity}
\title{find new linearly proximal genomic features in light of rearrangements}
\usage{
proximity(query, subject, verbose = F, mc.cores = 1, max.dist = 1e+06)
}
\arguments{
\item{query}{GRanges of "intervals of interest" eg regulatory elements}

\item{subject}{GRanges of "intervals of interest" eg genes}

\item{verbose}{logical flag}

\item{mc.cores}{how many cores (default 1)}

\item{max.dist}{maximum genomic distance to store and compute (1MB by default) should roughly
be the maximum distance at which biological interactions may occur}

\item{ra}{GRangesList of junctions (each a length 2 GRanges, similar to input to karyograph)}

\item{jab}{existing JaBbA object (overrides ra input)}
}
\value{
list of n x m sparse distance matrices:
$ra = subject-query distance in the rearranged genome for all loci < max.dist in tumor genome
$wt = subject-query distance in the reference genome for all loci < max.dist in tumor genome
$rel = subject-query distance in ra relative to wild type for above loci
}
\description{
Takes a set of n \code{query} elements (\code{GRanges} object, e.g. genes) and determines
their proximity to m \code{subject} elements (\code{GRanges} object, e.g. regulatory elements)
subject to set of rearrangement adjacencies (\code{junctions}). This redefined genomic
distance is computed by \code{gGraph$dist} and is a lower bound when all junctions utilized
on the shortest paths are in cis (on the same molecule).
}
\details{
Values x_ij in these matrices should be interpreted with a 1e-9 offset to yield the
actual value y_ij, i.e. y_ij = x_ij-1e-9, x_ij>0, y_ij = NA otherwise (allows for sparse
encoding of giant matrices)
}
