% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventCallers.R
\name{qrp}
\alias{qrp}
\title{Finds (quasi) reciprocal pairs of junctions.  Very related to chromoplexy or tic "cycles", but with
exactly two junctions.}
\usage{
qrp(
  gg,
  max.insert = 50000,
  min.cushion = 1e+06,
  min.span = 1e+06,
  ignore.loose.ends = TRUE,
  ignore.small.dups = TRUE,
  ignore.small.dels = TRUE,
  max.small = 50000,
  mark = FALSE,
  mark.col = "purple"
)
}
\arguments{
\item{gg}{gGraph}

\item{max.insert}{max insert to consider in a templated insertion (5e4)}

\item{min.cushion}{minimum cushion between a TIC junction and any other nearby event (to ensure "clean" events), the bigger the cushion, the cleaner the calls}

\item{min.span}{min span for a TIC junction (1e6)}

\item{ignore.loose.ends}{logical flag (FALSE) determining whether we ignore loose ends when filtering on min.cushion}

\item{ignore.small.dups}{logical flag (FALSE) determining whether we ignore small dups when filtering on min.cushion}

\item{ignore.small.dels}{logical flag (FALSE) determining whether we ignore small dels when filtering on min.cushion}

\item{max.small}{threshold for calling a local dup or del "small" 1e4}
}
\value{
gGraph with $meta annotated with gWalks corresponding to tic and tip and nodes and edges labeled with 'p1' through 'pk' for all k templated insertion paths and 'c1' through 'ck' for all k templated insertion cycles
}
\description{
Finds (quasi) reciprocal pairs of junctions.  Very related to chromoplexy or tic "cycles", but with
exactly two junctions.
}
