% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ra.duplicated}
\alias{ra.duplicated}
\title{ra.duplicated}
\usage{
ra.duplicated(grl, pad = 500, ignore.strand = FALSE)
}
\arguments{
\item{grl}{GRangesList representing rearrangements to be merged}

\item{pad}{non-negative integer specifying padding}

\item{ignore.strand}{whether to ignore strand (implies all strand information will be ignored, use at your own risk)}
}
\value{
\code{GRangesList} of merged junctions with meta data fields specifying which of the inputs each outputted junction was "seen.by"
}
\description{
Show if junctions are Deduplicated

Determines overlaps between two or more piles of rearrangement junctions (as named or numbered arguments) +/- padding
and will merge those that overlap into single junctions in the output, and then keep track for each output junction which
of the input junctions it was "seen in" using logical flag  meta data fields prefixed by "seen.by." and then the argument name
(or "seen.by.ra" and the argument number)
}
\author{
Xiaotong Yao
}
