% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R, R/misc_utils.R
\name{rand.string}
\alias{rand.string}
\title{make a random string}
\usage{
rand.string(n = 1, length = 12)

rand.string(n = 1, length = 12)
}
\value{
random string

random string
}
\description{
make a random string

make a random string
}
\author{
Someone from Stackoverflow

Someone from Stackoverflow
}
