% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_cmap}
\alias{read_cmap}
\title{read_cmap}
\usage{
read_cmap(path, gr = TRUE, seqlevels = NULL)
}
\arguments{
\item{path}{path to cmap file}
}
\description{
Reads cmap as GRanges
using
https://bionanogenomics.com/wp-content/uploads/2017/03/30039-CMAP-File-Format-Specification-Sheet.pdf
}
\author{
Marcin Imielinski
}
