% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{read_gencode}
\alias{read_gencode}
\title{read_gencode}
\usage{
read_gencode(con = Sys.getenv("DEFAULT_GENE_ANNOTATION"), type = NULL,
  by = NULL)
}
\arguments{
\item{con}{a path, URL, or \code{GFFFile} object pointing to gene annotation data}

\item{type}{the value}
}
\description{
reads GENCODE file dump from .rds of GRanges.

If that file doesn't exist
}
\author{
Marcin Imielinski
}
