% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{read_vcf}
\alias{read_vcf}
\title{read_vcf: utility function to read VCF into GRanges object}
\usage{
read_vcf(fn, gr = NULL, hg = "hg19", geno = NULL, swap.header = NULL,
  verbose = FALSE, add.path = FALSE, tmp.dir = "~/temp/.tmpvcf", ...)
}
\description{
read_vcf: utility function to read VCF into GRanges object
}

