% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventCallers.R
\name{reciprocal}
\alias{reciprocal}
\title{Get reciprical connected junctions}
\usage{
reciprocal(gg, thresh = 5e+05, max.small = 10000)
}
\arguments{
\item{gg}{gGraph}

\item{thresh}{threshold for edge clusters. Default: 5e5}

\item{max.small}{max small threshold for edge clusters. Default: 1e4}
}
\value{
gGraph with $ecluster marking on nodes and edges labeling unique reciprocal events
}
\description{
Identifies reciprocally connected junctions,
i.e. breakends from non-identical junctions that are "linked"
by an inter-breakpoint distance less than a given threshold.
Edges and nodes are marked by the "ecluster" metadata field
}
\details{
Reciprocal junctions are junctions with breakends that are mutually adjacent and opposite.

i.e. breakends from non-identical junctions that are "linked"
by an inter-breakpoint distance less than a given threshold.
Edges and nodes are marked by the "ecluster" metadata field
}
