% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R, R/misc_utils.R
\name{rep_len2}
\alias{rep_len2}
\title{recycle vector along length OR nrow of object}
\usage{
rep_len2(x, objalong, uselen = TRUE)

rep_len2(x, objalong, uselen = TRUE)
}
\arguments{
\item{x}{data}

\item{objalong}{any object to recycle x along if uselen = TRUE, or an actual integer value if uselen = FALSE}
}
\value{
vector

vector
}
\description{
recycle vector along length OR nrow of object

recycle vector along length OR nrow of object
}
\author{
Kevin Hadi

Kevin Hadi
}
