% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{rev.comp}
\alias{rev.comp}
\title{rev.comp
Return the reverse complement of a given GRanges}
\usage{
rev.comp(gr)
}
\arguments{
\item{gr}{GRanges with strand specified}
}
\value{
reverse complement of the input
}
\description{
rev.comp
Return the reverse complement of a given GRanges
}
