% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R, R/misc_utils.R
\name{rleseq}
\alias{rleseq}
\title{numbers up within repeating elements of a vector}
\usage{
rleseq(
  ...,
  clump = TRUE,
  recurs = FALSE,
  na.clump = TRUE,
  na.ignore = FALSE,
  sep = paste0(" ", rand.string(length = 6), " "),
  use.data.table = TRUE
)

rleseq(
  ...,
  clump = TRUE,
  recurs = FALSE,
  na.clump = TRUE,
  na.ignore = FALSE,
  sep = paste0(" ", rand.string(length = 6), " "),
  use.data.table = TRUE
)
}
\arguments{
\item{...}{Vector(s) to identify with unique id and a running id within each unique id}

\item{clump}{a logical specifying if duplicates are to be counted together}

\item{recurs}{a logical that is meant to only be set by the function when using clump = TRUE}
}
\value{
a list of idx and seq

a list of idx and seq
}
\description{
returns unique id within each unique element of a vector or set of provided vectors
and also a running id within each unique element

returns unique id within each unique element of a vector or set of provided vectors
and also a running id within each unique element
}
\author{
Kevin Hadi
}
