% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gurobi.R
\name{run_gurobi}
\alias{run_gurobi}
\title{run_gurobi}
\usage{
run_gurobi(
  cvec = NULL,
  Amat = NULL,
  bvec = NULL,
  Qmat = NULL,
  lb = NULL,
  ub = NULL,
  sense = NULL,
  vtype = NULL,
  objsense = "min",
  control = list(epgap = 0.01, tilim = 360, trace = 2),
  threads = 16
)
}
\description{
Rcplex2-like wrapper for calling Gurobi for MIP optimization
}
