% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R, R/misc_utils.R
\name{seq_along2}
\alias{seq_along2}
\title{seq along either row of table or length of vector}
\usage{
seq_along2(x)

seq_along2(x)
}
\arguments{
\item{x}{data}
}
\value{
vector

vector
}
\description{
seq along either row of table or length of vector

seq along either row of table or length of vector
}
\author{
Kevin Hadi
}
