% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{seqinfo}
\alias{seqinfo}
\title{seqinfo}
\usage{
\S4method{seqinfo}{gGraph}(x)

\S4method{seqinfo}{gNode}(x)

\S4method{seqinfo}{gEdge}(x)

\S4method{seqinfo}{Junction}(x)

\S4method{seqinfo}{gWalk}(x)
}
\arguments{
\item{x}{a gWalk}
}
\value{
the seqinfo of this graph

the seqinfo of this gNode

the seqinfo of this gEdge

the seqinfo of this Junction

the seqinfo of this gWalk
}
\description{
seqinfo

seqinfo

seqinfo

seqinfo

seqinfo
}
