% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{seqlengths}
\alias{seqlengths}
\title{seqlengths}
\usage{
\S4method{seqlengths}{gGraph}(x)

\S4method{seqlengths}{gNode}(x)

\S4method{seqlengths}{gEdge}(x)

\S4method{seqlengths}{Junction}(x)

\S4method{seqlengths}{gWalk}(x)
}
\arguments{
\item{x}{a gWalk}
}
\value{
the seqlengths of this graph

the seqlengths of this gNode

the seqlengths of this gEdge

the seqlengths of this Junction

the seqlengths of this gWalk
}
\description{
seqlengths

seqlengths

seqlengths

seqlengths

seqlengths
}
