% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R, R/misc_utils.R
\name{setColnames}
\alias{setColnames}
\title{convenience function to set column names}
\usage{
setColnames(object = nm, nm = NULL, pattern = NULL, replacement = "")

setColnames(object = nm, nm = NULL, pattern = NULL, replacement = "")
}
\arguments{
\item{object}{tabled object}

\item{nm}{names of the new columns}
}
\value{
colnamed object

colnamed object
}
\description{
convenience function to set column names

convenience function to set column names
}
\author{
Kevin Hadi

Kevin Hadi
}
