% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R, R/misc_utils.R
\name{setcols}
\alias{setcols}
\title{convenience function to set columns

sets columns of an object}
\usage{
setcols(dt, old, new)

setcols(dt, old, new)
}
\arguments{
\item{dt}{data frame/table or matrix}

\item{old}{integer or character or logical vector corresponding to current colnames in dt}

\item{new}{character vector for new column names}
}
\value{
colnamed object

colnamed object
}
\description{
convenience function to set columns

sets columns of an object

convenience function to set columns

sets columns of an object
}
\author{
Kevin Hadi
}
