% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{setdiff}
\alias{setdiff}
\title{setdiff.gNode}
\usage{
\S4method{setdiff}{gNode,gNode}(x, y)

\S4method{setdiff}{gEdge,gEdge}(x, y)

\S4method{setdiff}{Junction,Junction}(x, y, pad = 0, ...)
}
\arguments{
\item{x}{a Junction Object}

\item{y}{a Junction Object}
}
\value{
new gNode Object containing the difference between x and y

new gEdge containing the difference between x and y

new Junction Object containing the difference between x and y
}
\description{
Returns a new gNode which is the difference between x and y (node.id's).
All arguments must point at the same graph or an error will be thrown.

Returns a new gEdge which is the difference between x and y (edge.id's).
All arguments must point at the same graph or an error will be thrown.

Returns a new Junction Object which is the difference between x and y.
}
\author{
Joe DeRose

Joe DeRose

Rick Mortensen
## @exportMethod setdiff
}
