% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventCallers.R
\name{simple}
\alias{simple}
\title{Simple Event Calling}
\usage{
simple(
  gg,
  reciprocal.thresh = 10000,
  tra.pad = 1e+06,
  mark = TRUE,
  mark.col = "purple"
)
}
\arguments{
\item{gg}{gGraph, must have 'cn' node and edge annotation}

\item{reciprocal.thresh}{reciprocal threshold added to alt.shadows to find
translocations
Default: 1e4}

\item{tra.pad}{translocation padding to add on when finding locations
Default: 1e6}

\item{mark}{logical flag specifying whether to mark fold backs
Default: TRUE}

\item{mark.col}{character specifying colors to mark graph with
Default: 'purple'}
}
\value{
gGraph object containing labeling the putative event
}
\description{
Call simple events in gGraph
}
\details{
simple event calling includes 3 events: inversion, inverted
duplication, and translocation.

For how to run this function:
\href{http://mskilab.com/gGnome/tutorial.html#Simple}{Simple}
}
