% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{sort_snodes}
\alias{sort_snodes}
\title{sort a signed nodelist unambiguously, for use in hashing}
\usage{
sort_snodes(nodelist, circ = NULL)
}
\arguments{
\item{nodelist}{list of signed integers pointing to nodes in some gGraph}

\item{circ}{boolean vector indicating which of the node vectors is circular}
}
\description{
sort a signed nodelist unambiguously, for use in hashing
}
