% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{split.gGraph}
\alias{split.gGraph}
\title{split.gGraph}
\usage{
\method{split}{gGraph}(gg, by = "parent.graph")
}
\arguments{
\item{by}{field name to split by, default "parent.graph"}
}
\value{
A new gGraph object that is the union of the nodes and edges in the input gGraphs
}
\description{
Split gGraph by a node metadata field
}
\author{
Marcin Imielinski
}
