% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spmelt}
\alias{spmelt}
\title{spMelt}
\usage{
spmelt(A, baseval = 0)
}
\arguments{
\item{A}{}
}
\value{
data.table of all non
}
\description{
Melts sparse matrix into data.table
}
\author{
Marcin Imielinski
}
