% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{[.gEdge}
\alias{[.gEdge}
\title{gEdge}
\usage{
\method{[}{gEdge}(obj, i = NULL, from = NULL, to = NULL, with = TRUE, ...)
}
\arguments{
\item{obj}{This is the gEdge object to be subset}

\item{i}{integer, logical, or expression in gEdge metadata used to subset gEdge}
}
\value{
A new gEdges that contains only the given id's
}
\description{
Overloads subset operator for gEdge. Allows subsetting of gEdge via index or
sedge.id (as a key). Also, allows for data.table style queries on gEdge and
corresponding metadata.
}
\author{
Joe DeRose
}
