% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{[.gNode}
\alias{[.gNode}
\title{gNode}
\usage{
\method{[}{gNode}(obj, i = NULL, with = TRUE, ...)
}
\arguments{
\item{obj}{This is the gNode object to be subset}

\item{i}{integer, logical, or expression in gNode metadata used to subset gNodes}
}
\value{
A new gNode that contains only the given id's
}
\description{
Overloads subset operator for gNode. Allows subsetting of gNode via index or
snode.id (as a key). Also, allows for data.table style queries on gNode GRanges
and corresponding metadata.
}
\author{
Joe DeRose
}
