% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{[.gWalk}
\alias{[.gWalk}
\title{gWalk}
\usage{
\method{[}{gWalk}(obj, i = NULL, with = TRUE, ...)
}
\arguments{
\item{obj}{gWalk object This is the gWalk object to be subset}

\item{i}{integer, logical, or expression on gWalk metadata used to subset gWalk}
}
\value{
A new gWalk object that contains only the given subset of walks
}
\description{
Overloads subset operator for walks
}
