% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{[.Junction}
\alias{[.Junction}
\title{Junction}
\usage{
\method{[}{Junction}(obj, i, j, with = TRUE)
}
\arguments{
\item{obj}{Junction object This is the Junction object to be subset}

\item{i}{integer, logical, or expression in Junction metadata used to subset Junction}
}
\value{
A new Junction object that contains only the given id's
}
\description{
Overloads subset operator for Junction. Allows subsetting of Junction via index.
Also, allows for data.table style queries on Junction GRangesList and
corresponding metadata.
}
