% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R, R/gGraph.R
\name{[.junctions}
\alias{[.junctions}
\title{`[` -- subsetting a junctions object}
\usage{
\method{[}{junctions}(x, idx = NULL)

\method{[}{junctions}(x, idx = NULL)
}
\arguments{
\item{numeric}{vector of indices}

\item{numeric}{vector of indices}
}
\description{
`[` -- subsetting a junctions object

`[` -- subsetting a junctions object
}

