% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{sum.gGraph}
\alias{sum.gGraph}
\title{sum.gGraph}
\usage{
\method{sum}{gGraph}(..., by = NULL, na.rm = TRUE)
}
\arguments{
\item{...}{set of gGraph arguments (names will be incorporated as graph metadata $parent.graph)}

\item{y}{gGraph}
}
\value{
A new gGraph object that is the sum of the component gGraphs
}
\description{
Sums gGraphs using default aggregation operation for metadata (sum for numeric and paste(..., collapse = ',') for character, and c() for list)
}
