% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\docType{class}
\name{symsegs}
\alias{symsegs}
\title{nodes of a skew-symmetric genomic graph}
\description{
This is a special class of \code{GRanges} in which there defined a mapping function \code{hb}
that maps the pairs of two ssDNA that constitutes the dsDNA molecules such that at any time
ssDNA \code{s1} is present in an \code{symsegs} object, \code{s2=hb(s1)} must also be present.

The mapping funciton \code{hb} can be think of as getting the reverse complement of a ssDNA.
Nevertheless, the properties that sufficiently and necessarily define a unique ssDNA is not
merely its genomic ranges, but flexible and extendible by defining the property fields of
it. For example, we can have a type-\code{factor} field called "alleles" to represent each
different alleles even of the same genomic ranges. Obviously one can define an arbitrary number
of property fields on a \code{symsegs} object and the set of the elements is the Cartisan
product of all of them. To extend your imagination, we can also have "loose" for loose ends,
or "sample.id" for storing multiple samples in the same object. A property field must be
boolean, categorical, or nominal.
}
