% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R, R/misc_utils.R
\name{`\%nin\%`}
\alias{`\%nin\%`}
\alias{\%nin\%}
\title{not %in%

Not match}
\usage{
x \%nin\% table

x \%nin\% table
}
\arguments{
\item{x}{value to test}

\item{table}{table to test x against}
}
\description{
not %in%

Not match

not %in%

Not match
}
