% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transpose}
\alias{transpose}
\title{transpose}
\usage{
transpose(lst, ffun = list)
}
\description{
transpose using base R
}
\details{
Transpose a list without purrr or data.table
to avoid external dependencies. E.g. not all 
data.table::transpose versions are
fully compatible with lists
}
