% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{union}
\alias{union}
\title{union.gNode}
\usage{
\S4method{union}{gNode,gNode}(x, y, ...)

\S4method{union}{gEdge,gEdge}(x, y, ...)

\S4method{union}{Junction,Junction}(x, y)
}
\arguments{
\item{x}{a Junction Object}

\item{y}{a Junction Object}
}
\value{
new gNode Object containing the union of x and y

new gEdge containing the union of x and y

new Junction Object containing the union of x and y
}
\description{
Returns a new gNode which is the union of x and y (node.id's). All
arguments must point at the same graph or an error will be thrown.

Returns a new gEdge which is the union of x and y (edge.id's).
All arguments must point at the same graph or an error will be thrown.

Returns a new Junction Object which is the union of x and y.
}
\author{
Joe DeRose

Rick Mortensen
}
