% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{unique}
\alias{unique}
\alias{unique.Junction}
\title{unique.Junction}
\usage{
\method{unique}{Junction}(x, pad = 0, ignore.strand = FALSE)
}
\arguments{
\item{x}{a Junction Object}
}
\value{
new Junction Object containing the union of x and y
}
\description{
Returns the subset of Junction object that it is unique
}
\author{
Rick Mortensen
}
