% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\docType{data}
\name{walk2}
\alias{walk2}
\alias{bGraph}
\title{walk2}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
bGraph
}
\arguments{
\item{jab}{JaBbA object
#}
}
\value{
GRangesList of walks with copy number as field $cn, cyclic walks denoted as field $is.cycle == TRUE, and $wid (width) and $len (segment length) of walks as additional metadata
}
\description{
Computes greedy collection (i.e. assembly) of genome-wide walks (graphs and cycles) by finding shortest paths in JaBbA graph.
}
\details{
Descendant of gGraph class, where junction balance restraint must be met at all times
}
\keyword{datasets}
