% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{width}
\alias{width}
\title{width}
\usage{
\S4method{width}{gWalk}(x)

\S4method{width}{gNode}(x)
}
\arguments{
\item{x}{a gNode}
}
\value{
the width of each walk in this gWalk

the width of each gNode
}
\description{



}
