% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\docType{class}
\name{bGraph-class}
\alias{bGraph-class}
\alias{bGraph}
\title{junction-balanced graph}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
bGraph
}
\description{
Descendant of gGraph class, where junction balance restraint must be met at all times.
}
\keyword{datasets}
