% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\docType{data}
\name{bGraph}
\alias{bGraph}
\title{bGraph: junction-balanced graph
Descendant of gGraph class, where junction balance restraint must be met at all times.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
bGraph
}
\description{
bGraph: junction-balanced graph
Descendant of gGraph class, where junction balance restraint must be met at all times.
}
\keyword{datasets}

