% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R, R/gGraph.R
\name{components.gGraph}
\alias{components.gGraph}
\title{components -- same idea as igraph::components, returns a list of gGraph objects}
\usage{
\method{components}{gGraph}(gGraph)

\method{components}{gGraph}(gGraph)
}
\value{
a list of gGraph objects representing each partition of the input

a list of gGraph objects representing each partition of the input
}
\description{
components -- same idea as igraph::components, returns a list of gGraph objects

components -- same idea as igraph::components, returns a list of gGraph objects
}

