% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{dim}
\alias{dim}
\alias{dim.gGraph}
\title{dim}
\usage{
\method{dim}{gGraph}(gGraph)
}
\arguments{
\item{gGraph}{a \code{gGraph} object}
}
\value{
the number of nodes and edges in the gGraph
}
\description{
The number of nodes and edges in the gGraph
}
