% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R, R/misc_utils.R
\name{enframe_list}
\alias{enframe_list}
\title{data table-ize list elements and add name column}
\usage{
enframe_list(
  lst,
  name = "name",
  value = "value",
  as.data.table = TRUE,
  rbind = TRUE,
  mc.cores = 1
)

enframe_list(
  lst,
  name = "name",
  value = "value",
  as.data.table = TRUE,
  rbind = TRUE,
  mc.cores = 1
)
}
\description{



}
\author{
Kevin Hadi

Kevin Hadi
}
