% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventCallers.R
\name{events}
\alias{events}
\title{Shortcut to call all simple and complex event types on JaBbA graph using standard settings on all event callers.}
\usage{
events(gg, verbose = TRUE, mark = FALSE, QRP = FALSE)
}
\arguments{
\item{gg}{gGraph}
}
\value{
gGraph with nodes and edges annotated with complex events in their node and edge metadata and in the graph meta data field $events
}
\description{
Shortcut to call all simple and complex event types on JaBbA graph using standard settings on all event callers.
}
