% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventCallers.R
\name{fusions}
\alias{fusions}
\title{find fusions}
\usage{
fusions(
  graph = NULL,
  gencode = NULL,
  genes = NULL,
  annotate.graph = TRUE,
  mc.cores = 1,
  verbose = FALSE
)
}
\arguments{
\item{graph}{input gGraph}

\item{gencode}{GFF containing gene boundaries and exons, in similar format to
https://www.gencodegenes.org/}

\item{genes}{set of genes to pass for fusions.}

\item{annotate.graph}{Annotate the graph generated Default: True}

\item{mc.cores}{number of cores to run. Default: 1}

\item{verbose}{output verbose argument to function Default: False}
}
\value{
gWalks of gene fusions annotated with frame and gene(s)
}
\description{
annotates all gene fusions in gGraph relative to cds definitions

cds = gencode cds GRanges gff3 / gtf or GRangesList the latter (converted via rtracklayer::import)
which has fields $exon_number

"gene_name" GRangesList meta data field is used in annotation and in not
creating "splice" fusionsthat arise from different transcripts of the same gene.
}
\details{
For more info please follow the protein fusions analysis in the gGnome Tutorial:

\href{http://mskilab.com/gGnome/tutorial.html#Protein_fusions}{Protein fusions}
}
