% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\docType{data}
\name{gGraph}
\alias{gGraph}
\title{gGraph}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
gGraph
}
\description{
the central class for rearrangement graphs
}
\details{
#' @import BSgenome
}
\keyword{datasets}
