% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{gNode}
\alias{gNode}
\title{gNode}
\description{
gNode object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-gNode-new}{\code{gNode$new()}}
\item \href{#method-gNode-mark}{\code{gNode$mark()}}
\item \href{#method-gNode-rep}{\code{gNode$rep()}}
\item \href{#method-gNode-swap}{\code{gNode$swap()}}
\item \href{#method-gNode-subset}{\code{gNode$subset()}}
\item \href{#method-gNode-eval}{\code{gNode$eval()}}
\item \href{#method-gNode-clusters}{\code{gNode$clusters()}}
\item \href{#method-gNode-ego}{\code{gNode$ego()}}
\item \href{#method-gNode-print}{\code{gNode$print()}}
\item \href{#method-gNode-loose.degree}{\code{gNode$loose.degree()}}
\item \href{#method-gNode-clone}{\code{gNode$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gNode-new"></a>}}
\if{latex}{\out{\hypertarget{method-gNode-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gNode$new(snode.id = NULL, graph)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gNode-mark"></a>}}
\if{latex}{\out{\hypertarget{method-gNode-mark}{}}}
\subsection{Method \code{mark()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gNode$mark(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gNode-rep"></a>}}
\if{latex}{\out{\hypertarget{method-gNode-rep}{}}}
\subsection{Method \code{rep()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gNode$rep(times)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gNode-swap"></a>}}
\if{latex}{\out{\hypertarget{method-gNode-swap}{}}}
\subsection{Method \code{swap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gNode$swap(replacement)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gNode-subset"></a>}}
\if{latex}{\out{\hypertarget{method-gNode-subset}{}}}
\subsection{Method \code{subset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gNode$subset(i)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gNode-eval"></a>}}
\if{latex}{\out{\hypertarget{method-gNode-eval}{}}}
\subsection{Method \code{eval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gNode$eval(x, right = TRUE, all = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gNode-clusters"></a>}}
\if{latex}{\out{\hypertarget{method-gNode-clusters}{}}}
\subsection{Method \code{clusters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gNode$clusters(mode = "weak")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gNode-ego"></a>}}
\if{latex}{\out{\hypertarget{method-gNode-ego}{}}}
\subsection{Method \code{ego()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gNode$ego(order = 0, mindist = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gNode-print"></a>}}
\if{latex}{\out{\hypertarget{method-gNode-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gNode$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gNode-loose.degree"></a>}}
\if{latex}{\out{\hypertarget{method-gNode-loose.degree}{}}}
\subsection{Method \code{loose.degree()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gNode$loose.degree(orientation)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gNode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-gNode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
