% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcplex.R, R/eventCallers.R
\name{Rcplex2}
\alias{Rcplex2}
\alias{simple}
\title{Rcplex2}
\usage{
Rcplex2(
  cvec,
  Amat,
  bvec,
  Qmat = NULL,
  lb = 0,
  ub = Inf,
  control = list(),
  objsense = c("min", "max"),
  sense = "L",
  vtype = NULL,
  n = 1,
  tuning = FALSE
)

simple(
  gg,
  reciprocal.thresh = 10000,
  tra.pad = 1e+06,
  mark = TRUE,
  mark.col = "purple"
)
}
\arguments{
\item{gg}{gGraph, must have 'cn' node and edge annotation}

\item{mark}{logical flag specifying whether to mark fold backs (TRUE)}

\item{mark.col}{character specifying colors to mark graph with ('purple')}

\item{min.foldbacks}{minimal number of foldbacks that makes a bfb}

\item{amp.thresh}{amplification threshold for a BFB}

\item{cn.bfb}{minimal number of copies that the highest copy foldback in the bfb needs to have}
}
\value{
gGraph object containing labeling the putative event
}
\description{
Modification of Rcplex which takes in mipcontrol parameters.

Call simple events in gGraph
}
