% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\docType{class}
\name{junctions-class}
\alias{junctions-class}
\alias{junctions}
\title{junctions
S4 wrapper around GRangesList to store junction info}
\description{
junctions
S4 wrapper around GRangesList to store junction info
}
