% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{length}
\alias{length}
\alias{length.gNode}
\alias{length.gEdge}
\alias{length.Junction}
\alias{length.gGraph}
\alias{length.gWalk}
\title{length.gNode}
\usage{
\method{length}{gNode}(x)

\method{length}{gEdge}(x)

\method{length}{Junction}(Junction)

\method{length}{gGraph}(x)

\method{length}{gWalk}(x)

\method{length}{gNode}(x)

\method{length}{gEdge}(x)
}
\arguments{
\item{Junction}{a Junction Object}

\item{gNode}{a gEdge object}

\item{gWalk}{a \code{gWalk} object}

\item{gGraph}{a \code{gEdge} object}
}
\value{
the number of nodes in the gNode Object

the number of edge pairs in the gEdge Object

the number of junctions in the Junction Object

the number of nodes in the gGraph

the number of nodes in the gWalk

the number of nodes in the gNode

the number of edges in the gEdge
}
\description{
The number of nodes in the gNode Object

The number of edge pairs in the gEdge Object

The number of junctions in this Junction Object

The number of nodes in the gGraph

The number of walks in the gWalk

The number of nodes in the gNode

The number of edges in the gEdge
}
\author{
Joe DeRose

Joe DeRose
}
