% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{lengths}
\alias{lengths}
\alias{lengths.gWalk}
\title{lengths}
\usage{
lengths.gWalk(gWalk, use.names = FALSE)

\S4method{lengths}{gWalk}(x)
}
\arguments{
\item{gWalk}{a \code{gWalk} object}

\item{x}{a gWalk object}
}
\value{
the number of nodes in the gWalk

the number of nodes in each gWalk
}
\description{
A vector of walk lengths associated with this walk


}
