% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{lengths}
\alias{lengths}
\alias{lengths.gWalk}
\title{lengths}
\usage{
lengths(x, use.names = TRUE)

lengths.gWalk(x, use.names = FALSE)

\S4method{lengths}{gWalk}(x)
}
\arguments{
\item{x}{a gWalk object}

\item{gWalk}{a \code{gWalk} object}
}
\value{
the number of nodes per walk in the gWalk

the number of nodes per walk in the gWalk

the number of nodes in each gWalk
}
\description{
establish s3 method for lengths

A vector of walk lengths associated with this walk


}
