% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventCallers.R
\name{microhomology}
\alias{microhomology}
\title{get microhomology}
\usage{
microhomology(gg, hg)
}
\arguments{
\item{gg}{gGraph or Junctions}

\item{hg}{DNAStringSet or path to reference fasta}
}
\value{
gGraph with $pyrgo marking on nodes and edges labeling unique "events"
}
\description{
Computes microhomology at 5bp, 10bp, 50bp, and 100bp windows
 around ALT junctions of input gGraph (or Junction object)
gg and adds these as an edge annotation to the appropriate edges.

Requires Biostrings
}
