% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventCallers.R
\name{qrp}
\alias{qrp}
\title{Finds quasi-reciprocal pairs of junctions}
\usage{
qrp(
  gg,
  thresh = 1e+06,
  max.small = 1e+05,
  breakend_pairing = c("strict", "one_to_one", "loose"),
  mark = TRUE,
  mark.col = "purple"
)
}
\arguments{
\item{gg}{gGraph}

\item{thresh}{integer - maximal size of bridge}

\item{max.small}{integer indicating maximum size of candidate SVs for clustering}

\item{breakend_pairing}{"strict": can only be "monogamously" matched to one breakend and if the nearest breakend is of the wrong orientation, it is thrown out; "one_to_one:" breakends can only be coupled to a single "monogamous" match but without considering breakends of the wrong orientation. If breakend B is nearest to C, B will only be matched to C. If A's nearest breakend is B but is further away than C, A will not be matched to B. "loose": the nearest breakend in the correct orientation under the threshold is considered. The same as one_to_one except A will be matched to B, while B will be matched to C.}

\item{mark}{logical, mark the edges and nodes in color specified by mark.col}

\item{marck.col}{character, color to mark nodes/edges involved in any QRP}
}
\value{
gGraph with $ecluster marking on nodes and edges labeling unique reciprocal events
}
\description{
Finds quasi-reciprocal pairs of junctions
}
\author{
Kevin Hadi
}
