% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{ra.duplicated}
\alias{ra.duplicated}
\title{marking duplicated junctions as \code{TRUE}}
\usage{
ra.duplicated(junc)
}
\description{
Similar to \code{duplicated}, will return \code{logical} of the same length as the input,
where the \code{TRUE} denotes junctions that have been found in previous part of the input.
}
