% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{ra_breaks}
\alias{ra_breaks}
\title{ra_breaks: utility function to read junction data from various common formats}
\usage{
ra_breaks(rafile, keep.features = T, seqlengths = hg_seqlengths(),
  chr.convert = T, snowman = FALSE, swap.header = NULL,
  breakpointer = FALSE, seqlevels = NULL, force.bnd = FALSE, skip = NA,
  get.loose = FALSE)
}
\description{
ra_breaks: utility function to read junction data from various common formats
}

