% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{refresh}
\alias{refresh}
\title{refresh}
\usage{
refresh(x)

\S4method{refresh}{gGraph}(x)

\S4method{refresh}{gWalk}(x)
}
\arguments{
\item{Junction}{object}

\item{gGraph}{object}

\item{gWalk}{object}
}
\value{
Junction object

gGraph object

gWalk object
}
\description{
Updates Junction object to reflect changes in source code

Updates gGraph object to reflect changes in source code

Updates gWalk object
}
