% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setxor}
\alias{setxor}
\title{setxor}
\usage{
setxor(A, B)
}
\arguments{
\item{A}{vector specifying set A}

\item{B}{vector specifying set B}
}
\value{
elements in A or B that are not in the intersection of A and B
}
\description{
setxor
}
\author{
Marcin Imielinski
}
