% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gGnome.R
\name{setxor}
\alias{setxor}
\title{XOR operation on sets}
\usage{
setxor(A, B)
}
\arguments{
\item{A}{set A}

\item{B}{set B}
}
\value{
The set of elements belong to either A or B, but not both.
}
\description{
XOR operation on sets
}
\author{
Marcin Imielinski
}
