% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vaggregate}
\alias{vaggregate}
\title{\code{stats::aggregate}, but returns vector}
\usage{
vaggregate(...)
}
\arguments{
\item{...}{arguments to aggregate}
}
\value{
named vector indexed by levels of "by"
}
\description{
Same as \code{stats::aggregate} except returns named vector
with names as first column of output and values as second

Note: there is no need to ever use aggregate or vaggregate, just switch to data.table
}
\author{
Marcin Imielinski
}
\keyword{internal}
